OK so the work on assessment 2 has begun. README.md now details the task. I've had a crack at the first bit, specifically kuzu.py. I think it works ok, but it can be improved. At least it's a start. 2 weeks to deadline. Damn I'm tired.
The marks for quiz 1 and quiz 2 came out. I'm surprised i did better than expected. 76% for quiz 1 and 80% for quiz 2. It's not great, but it's a lot better than I expected. I was seriously prepared for maybe 50%. It seems that I have an unfortunate tendency to seriously doubt myself in this domain. I guess its down to experience, as this field is VERY different to what I'm used to.